Added node creation on drag and actions.
Dragging a connection into empty space was aborted up until now. Today
it started creating a new node and connecting to it.

Secondly, it is possible now to create actions (arrows showing shift in
time). You can do that by dragging from a circle on the right of the
node.