Update to jsplumb 1.7.4.
This jsplumb got a before drag interceptor, so we can now set the scope
using the source endpoint, which is a big improvement, because we do not
rely anymore on internals and hacking.

Kind of.

Also, using the data we got a pretty robust mechanism of detecting
double connections going through listeners.