improve the image generation.
My big problem was that the white shadow of node names was generated
with blur, and it simply was not working (it looked nice on chrome,
terrible in ImageMagick, and was invisible in Firefox).

I do not want the code to be pretty anymore. If those filters do not
work, I will just add the text twice - one as a blurred white, and the
second time for normal purpose.